Move en/disable_numba to utils
