+++ blpdm/lpd.py
@@ -7,7 +7,6 @@ efficiently (hopefully) with the help of numba
 # in order to pass in dicts as args for numba fn
 # need to use their special dict type and specify the types of the varibles
 #   https://numba.pydata.org/numba-doc/dev/reference/pysupported.html#typed-dict
-import importlib
 import os
 import sys
 import warnings
@@ -20,45 +19,10 @@ from numba import njit
 from numba import prange
 
 
-
-def disable_numba():
-    """Tell numba to disable JIT.
-
-    refs
-    ----
-    * disabling JIT: https://numba.pydata.org/numba-doc/latest/user/troubleshoot.html#disabling-jit-compilation
-    * reloading config: https://github.com/numba/numba/blob/868b8e3e8d034dac0440b75ca31595e07f632d27/numba/core/config.py#L369
-    """
-    os.environ.update({'NUMBA_DISABLE_JIT': str(1)})
-    numba.config.reload_config()
-    assert numba.config.DISABLE_JIT == 1  # pylint: disable=no-member
-
-    # note: reloading numba here does not change lpd's functions as imported by another module
-    #   like the model does
-    # until the module importing lpd reloads lpd (`importlib.reload(lpd)`)
-    # so the following does not have the desired effect
-    #
-    # importlib.reload(numba)
-    # from numba import jit, njit, prange
-
-
-def enable_numba():
-    """Tell numba to enable JIT.
-
-    see refs in `disable_numba`
-    """
-    # by default (when numba is loaded normally), this env var is not set, so remove it
-    os.environ.pop('NUMBA_DISABLE_JIT', default=None)  # avoid error
-    numba.config.reload_config()
-    assert numba.config.DISABLE_JIT == 0  # pylint: disable=no-member
-
-
 @njit
 def _calc_fd_params_above_canopy(pos, p):
-    """
-
-    tau's, dtau's, etc.
-
+    """Calculate the dispersion parameters above the canopy.
+    (tau's, dtau's, etc.)
     """
     z = pos[2]
